This is the initial commit of Sire to the google code repository. This is actually revision 831 of my personal svn repository.
Some extra files have been removed to keep the repository size down. I will eventually place the complete dump of my own repository online once google code allows file hosting. (its a 13MB bz2 file). If you want the original files before then, then please email me (chryswoods@gmail.com) and I will be happy to send you a copy.

Sire has been in development now for over a year and will be very powerful. However, it is still in an early developmental stage and is not ready for use. If you are interested in getting or developing sire then please email me, or look for the links for the sire mailing lists.

==========================================================================================
r831 | chris | 2006-07-25 21:55:38 +0100 (Tue, 25 Jul 2006) | 5 lines

I have fixed some bugs in the assignment of parameters and now the sb2 example works a little more (test/SireDB/assign_sb2.py). It even assigns the atom parameters for the perturbed state. However, this has highlighted a major logic error on my part. At the moment I am using parameter number '0' to signify dummy and not-found. This is a problem, as when the user sets the parameter 0, the db reports it as not found and will then try to find a better parameter (hence why sb2 perturbed state has a total charge of 0.315 - the three supposedly dummy hydrogens have been assigned reference state parameters, with charges of 0.105 each). I will need to change the code so it uses 0 for dummy, and -1 for not found. Such a change is best not attempted at 10pm - it would lead to too many tired bugs...

Otherwise the code compiles, links and appears to run well :-)