I have now updated CutGroup and made it into a class that holds a QVector of AtomInfo objects, together with a CoordGroup. It is now much more efficient to change coordinates, and to separate the coordinates from the atoms.
I have decided that CutGroup will not index its contents - it just contains a collection of atoms. The MoleculeInfo object is responsible for indexing a CutGroup, and allows mapping of an AtomIndex to a CGAtomID. There is no need for this functionality (or data) to be present in CutGroup as well.

Code is broken (still need to merge AABox and new CutGroup into the rest of the code)