I have done it - the python modules now compile and work via cmake on redhat. The problem was all caused by a missing symbolic link! Redhat had libboost_python.so.1, but did not have the symbolic link to libboost_python.so
This meant that the linker linked each python module to libboost_python.a - the static library! This caused a headache, as it probably screwed up the registry of types across the modules. While I had resolved most of the problems by separating the wrappers from the module definition, it didn't solve everything.

My solution (or should I say hack!) is to add the symbolic link from /usr/lib/libboost_python.so.1 to /usr/lib/libboost_python.so manually, and to then add a configure test to the CMakeLists.txt file that checks to see if the shared library has been found (if not, it dies with an error).