SireMol now compiles and links! I am finally beginning to make progress back to a working piece of code :-)
I updated CoordGroup so that it is a complete const-class - to edit it you must now use CoordGroupEditor.
I have also updated CoordGroupEditor so that it knows when the coordinates have been changed so that
it doesn't recalculate the bounding box unnecessarily. This helps when the coordinates are
copied directly from another CoordGroup.

I have added in all of the MoleculeData::setCoordinates functions. All of this now needs to be tested, but only
once I have updated the rest of the code so that I can recompile everything.