I have added lots of code to calculate CLJ energies, and InterCLJFF and CLJFF are now nearly complete (well, except for me working out exactly how to store molecules within a forcefield, but that is a problem for another day!).
Some of the functionality of InterCLJFF is missing, and it is not yet possible to add any molecules to this forcefield (quite important requirement, you'd think...!), and the python wrappers need writing, but otherwise, it all compiles and runs.

I've made some changes to SireCAS to allow operator= for Function and Symbol. This is probably a bad thing to do. I've realised that SireCAS should use DynamicSharedPtr rather than boost::shared_ptr, as this would simplify the lifetime management, allow operator= for all SireCAS objects, simplify serialisation (as would use DynamicSharedPtr serialisation) and would make the code more robust. This is a change for another day.

Code is working again. (probably not for long)