I've updated the code to allow successful compilation using gcc4 symbol visibility hiding. This reduces the size of the python wrapper libraries by about a third (e.g. SireMol goes down from 31MB to 18MB). The size of the wrappers is still ridiculous, but at least its a start...!
Also, Tip4PFF is now faster than Sire(493)! The old prototype took 560ms to calculate the energy, while Tip4PFF takes 540ms on my laptop. The correct InterCLJFF takes about 800ms, so is about 50-60% too slow, but at least now this is looking like an achievable goal.

I've also now filled in the missing functions for the recalulate delta energy. I've realised that it is best for the current configuration to be the new configuration, as then the forcefield will publically display only a single state. The old configuration is stored in the ChangedMolCLJInfo objects. I've also fixed a bug in the algorithm that would have missed the energy change of molecules that are removed.

The code compiles, links and runs well.