I am still sorting out the implementation of parallel processors. I have now almost fully fleshed out the design, and now need to write some tests and get it all to compile and link.
I have slightly changed the design so that now the forcefield calculation itself has been separated from the network model used (i.e. now I have a class hierarchy based on FFCalculator, which is used to calculate FF energies, and another, separate class hierarchy based on ProcessorBase that is used to represent the remote processor). This allows different ways of calculating a forcefield to be more easily combined with different network models, e.g. now MolproFF can use MolproCalculator to get the energy, and FFThreadProcessor to do this in a separate thread, or FFMPIProcessor to do this on a separate MPI node (not that I have FFMPIProcessor yet....)

Code is definitely broken.