The design of MoleculeGroups and ForceFields is now becoming clear - this does clean up the code in SystemData. Its not quite right yet, as I don't like what is happening with the ForceFields pointer in SystemData (which perhaps ought to be renamed SystemBase).
Also, I think that I have worked out the role of System - it is to be the checkpoint and editable front end. SystemData/Base has functions that may only be called during a normal move. An editing type of move will have to convert to System, make the change, then set from System, rather than rollback the checkpoint.

Code is very broken.