I now really have a strong handle on how the System, SystemData and SimSystem classes are working, and am making good progress with the SimSystem class in particular.
System is the user-facing class, which allows the user to construct and edit a System.

SystemData holds all metadata about a system, except for any forcefield data.

ForceFields holds all of the forcefield information, including expression information.

SimSystem is a noncopyable class that holds a simulated System, and synchronises SystemData and ForceFields throughout the simulation. SimSystem also does the work necessary to ensure that each move is atomic (state changes are complete)

ForceFields is a virtual class, allowing derived classes to spread the forcefields across different processors. ForceFields is the only virtual class, thus making things clean and nice.

Code is still broken.