After much painful debugging I have got the MolproFF working, and producing what I think is the right energy. The deadlocking problem was caused by the small output buffer of QProcess. Basically;
(1) Sire uses RPC to request an energy calculation
(2) Molpro performs the calculation, then writes lots of status output to stdout
(3) The QProcess stdout buffer becomes full
(4) Molpro blocks on a write, waiting for the stdout buffer to be cleared or read
(5) Sire is still blocked waiting for Molpro's response, so is unable to clear or read the buffer

My solution was to ignore stdout and to send the information I need to stderr (and to read that instead). It's not the best solution.

I think that QProcess may have gone through some changes between Qt 4.1 and Qt 4.2 as I didn't see this behaviour on my laptop (Qt 4.1) while I did no my desktop (Qt 4.2). Also, 4.2 introduces new input/output functions for QProcess, which may have lead to a change in code. I'll take a look at this at home when I can get back on my laptop. If this is not the case, then the fix I have introduced works only for Qt 4.2 (and currently Unix as well) so I will need to investigate another more robust solution.

Code compiles, links and runs well.