I've added the code to convert from Molpro energy units (hartrees) to Sire energy units (kcal mol-1) and have also added the origin energy (so that Molpro QM energies are of the same magnitude as the MM energies).
The Molpro QM energies appear to be very insensitive to moving the QM molecule - indeed, the difference in energy between just the QM molecule and the QM molecule surrounded by MM lattice points is only 0.0002 hartrees (~0.1 kcal mol-1). This seems really small to me, and is probably a result of an error on my part.

Code compiles, links and runs