I've updated the python wrapping to add a __copy__ function that uses the copy constructor (so that my classes now work with copy.copy).
I have wrapped up InterGroupLJFF, fixed its used of FFBase::Groups, and have updated the test/test.py script to test that a set of InterGroupLJFF forcefields return the same energy as a single InterLJFF forcefield, which itself gives the same energy as the LJ component of InterCLJFF! Also, 1000 moves on the combination of InterGroupLJFF forcefields takes the same time as 1000 moves on InterLJFF, so I am pretty pleased!

Code compiles, links and runs well.