I am still working on extending the ForceField API across to the entire program.
A key change now is that I have created PartialMolecule, which can represent any part of a molecule all the way from an Atom, through Residue to a Molecule, with everything in between. This has very much simplified the ForceField API, which is now beginning to look a little less oppressive than it did! 

The hierarchy of classes are;

FFBase forcefields (do the work)
    |
ForceField - provides implicit sharing and consistent interface
    |
FFCalculator - provides the server-side handle for a forcefield,
    |          used on the server side of a Processor
    |          (e.g. MolproCalculator). A calculator performs
    |          the work for the forcefield in the current thread
    |
FFWorker - provides the client-side handle for a forcefield
    |      and calls the forcefield via the FFCalculator
    |      (e.g. FFThreadWorker). A worker can run the work
    |      of the forcefield in a background or remote thread
    |
ForceFields - provides the interface to collections of forcefields,
              either interfaced directly via ForceField objects, or 
              interfaced on remote threads/processors via FFWorker objects.
              (e.g. ForceFields vs. ClusterForceFields)