I'm making good progress with SimSystem and SystemData. I've cleaned up the code so that now adding to the SystemData will map the molecule into the system box, then returning the result so that it can be used when adding to ForceFields.
I've made a small change to the behaviour of the FFBase::add() functions - now the whole molecule is changed whenever a difference is noticed when adding a molecule to an existing molecule. The reason for this is to prevent errors whereby a small part of a molecule with a totally different version is added - the user will use the selection to say what is being added, not what has changed, so using the selection would be a mistake in this case.

Code is broken.