I've fixed the bug that was causing what I thought was an infinite loop in water_mc_speed.py - I had not cleared the 'need_total_recalc' flag in InterGroupCoulombFF, so it was calculating the energy from scratch for each move!
I have writen new unit tests to discover the problem with MolproFF. Indoing so, I have added a skeleton unit test, to speed up the creation of new ones, and have also moved the addition of tests torun into SireUnitTest, so that they don't all need to be included in the siretest executable.

Using this, and some more debugging, I have fixed MolproFF which now produces the correct energies.

Code compiles, links and runs, and siretest completes without error. The code is finally working again (well, at least as well as it was before Easter!)