I've finished updating all of the code to use the new Molecules class. It seems to work quite well, and I like the new syntax and API of the class (it makes things a lot more straightforward).
I've been playing with changing molecules in the forcefields, and there is now a failure with the InterGroupLJFF - I think that the problem is that my change code does not take into account that the molecule may exist in more than one group at the same time. I will need to update all three of the InterGroup???FF classes to take account of this.

Code compiles, links and runs, and siretest has two errors, both related to InterGroupLJFF.