I'm now making good progress with the design and implementation of the new molecule classes. Yes, yet another new molecule class design! However, this time learns from all of the lessons before, and is more of a consolidation of the ideas and half-implemented bits of code that have arisen during the development of this project. The new molecule design now more formally splits classes between view classes (Molecule, Segment, CutGroup, Chain, Residue, Atom), container view classes (SegmentsInMol, CutGroupsInMol, ChainsInMol, ResiduesInMol, AtomsInMol and PartialMolecule), and action classes, which are implemented as templates on top of the view and container view classes, e.g. Mover<Molecule>, Editor<Residue>, Evaluator<ChainsInMol>, Selector<PartialMolecule>. I have to use templates so that the original view type is restored on a commit, which is required for the python wrappers (which don't have the same copy semantics as C++).
The new design allows code such as this to be written;

atoms = mol.atoms( ResName("ALA") + ResNum(35) )

for i in range(0, len(atoms)):
    atoms = atoms.move(i).translate( i*Vector(1,2,3) )

mol = atoms.molecule()

Note how the container view class (AtomsInMol) solves the problem of moving part of a molecule not
updating the rest of the objects in the container (e.g. the following code would be wrong)

atoms = [ list of atoms from the molecule ]

for i in range(0, len(atoms)):
    atoms[i] = atoms[i].move().translate( i*Vector(1,2,3) )

This is because even if each of the atoms in 'atoms' are a view of the same version of the molecule
before moving, they become independent and view different versions of the molecule after moving. 

Code is incredibly broken now!!!