I'm making good progress sorting out the api and implementation of the molecule container classes. I've decided that containers will not be const classes (unlike Molecule). The reason is that I don't want to have to copy the container (of potentially thousands of molecules) every time it is updated. The API as a const class was a bit messy.
Now I have made Molecules a non-const class, and have created the code that means that
it returns what has changed when it is modified. This will help the forcefield classes,
which will use the MolGroup class (that uses Molecule) to keep track of how things
change. This abstracts all molecule bookkeeping code into a single place (Molecules),
rather than as before, having it scattered across the forcefield classes.

Code is still broken (yes, it has been broken for most of 2007, and is still broken
now that we have moved into 2008 - it will however be fixed soon, as I have a lot
of production simulations planned that will use Sire this year!)