I've filled in the query API functions for MolGroupsBase. These are quite powerful. They will allow the user to write python code like;
for atom in system.selectAll( AtomName("CA") ):
    atom = atom.move().translate( (1*angstrom,0,0) )
    system.update(atom)

This code (obviously!) translates all of the atoms called "CA" in the system by 1 angstrom
to the right along the x axis. How cool is that! The user can also be more selective, e.g.

for atom in system.selectAll( MGName("protein") + AtomName("CA") ):

...or...

for atom in system.atoms( MGName("protein") + MolIdx(0) + AtomName("CA") ):

This last one moves all of the atoms called CA in the first protein in the "protein" group.

What is even more cool is that this API is the same for *all* MolGroupsBase derived objects,
i.e. System, ForceField and MoleculeGroups!!!

Code is broken.