I've further tested the new CoordGroup classes and they seem to work well. For two groups of 1000 molecules, where each molecule contains 3 groups of 4 atom (with random coords), the calculation of all interatomic distance pairs between groups takes 4.8s. It takes 2.0s to calculate distance^2 between all pairs. (thats 7000^2 pairs)
This compares to 5.9s and 2.9s for QVector< QVector<CoordGroup> > (which is not too bad).
This is a 20-30% speedup.

However, each Vector::distance2(Vector,Vector) calculation uses about 8 double operations.
This means that (ignoring loop counters and double comparisons) there are just 8*(7000^2)
operations in 2s, so just 4*(7000^2) double operations per second. This is only
100 MFLOPS (double), which is far from the 2-4 GFLOPS that my laptop should be capable of...

Code is broken