I've played around with trying to explicitly add SSE support to my Vector class, but without much success (I always end up with something that is slower than the compiler...).
I've also made the vector into a 32bit class, so it aligns better in arrays, and have
added some padding to the CoordGroup CGMemory which seems to have sped it up a little.

The fastest now is about 1.8s for distance^2, and 4.7s for distance().

(so nearly 3s is just sqrt!)

I was wrong in the last post, there are 12000 points in each group (1000 blocks
of 3 groups or 4 points). This means that it is 144M pair calculations, therefore
it is at least 8*144M operations. For dist2 this is about 0.65 GFLOPS, which is slightly
more respectable, while for distance it is 0.25 GFLOPS (although I don't know how many
operations are in a sqrt).

This is still a way from the theoretical value of 15-20 GFLOPS for my laptop...