I've now written SireFF:FF, the base class of all forcefields, and it compiles without error.
I've also moved MoleculeInfoBase over to using QUuid, and I will move everything that really
requires a unique ID over QUuid as well. This is because I really do need to ensure a 
unique UID across multiple machines, as well as time (as I can be loading and saving
files over time, and also passing data between nodes in a cluster). It is important that
once an objects has got a UID, that it keeps it!.

Code is broken