I've now written Intra2B2GFF, Intra2B2G3DFF and created IntraCLJFF and IntraGroupCLJFF (so now have the code to allow perturbation of parts of molecules, e.g. residues in a protein).
I've culled some files - particularly as I will now move all database files so that they
are separate to SireFF and SireMM (as I want the core runtime to not have to depend on QtSql
as it may not be available on compute nodes). I will therefore (eventually!) write
SireFFDB and SireMMDB libraries that will derive from SireDB and will contain the database
code necessary to parameterise the molecules in SireMM and SireFF.

With the code cleaned, SireFF now compiles and links, and all of SireMM compiles,
but there are some unresolved symbols.

Code is broken