I've written the code to calculate the bond, angle, dihedral, Urey-Bradley and stretch-stretch forces.
I've deliberately written readble and generic code, as this is for the generic InternalFF forcefield,
which is not expected to be the bottleneck in the simulation. If it becomes the bottleneck, then a
specific, speed-optimised version can be written, using the generic forcefield as a check and 
benchmark.

Code is broken