I've fixed the problem regarding the problem of not matching templates across shared library boundaries, and now Sire on the Mac is able to calculate the energy.
The problem was caused because GCC implements RTTI by comparing the pointer to the 
type_info object for the virtual class, rather than doing a string comparison of the 
name of the class (comparing type_info::name()). This breaks for a template class instantiated
in two different dynamically shared libraries (DSO) (or in a library and the main app) as G++ creates a copy of
the template class in each library, each with its own type_info object. These objects have
different pointers, so G++ reports that they are different objects and dynamic_cast fails.

The problem is seen when symbol hiding is turned on, as the type_info objects created for
each copy of the template class are not exported from the library. However, if the symbols
for the type_info objects were exported, then the linker would be able to see that they 
are the same, and would then fix the problem for us by merging the references. The solution,
therefore, is to ensure that the symbols for the type_info objects associated with 
polymorphic template classes are exported in all libraries where the template is 
instantiated.

On Linux, gcc will export all symbols by default. Only gcc >= 4 has the ability to hide
symbols, and this has to be deliberately turned on. However, if gcc symbol hiding is switched
on, then you will see this problem on Linux as well. (e.g. if -fvisibility-hidden is used)

On Mac and Windows symbol hiding is the default, so this is a problem on these systems.

One solution I saw in forums was to export *all* symbols in the libraries that instantiate
the template. While this would work, it is hardly ideal as you lose all the benefits of
using symbol hiding!

Fortunately, an easier solution is available. All you need to do is to mark the template
class itself as being visible, and this with then mark all of its functions and 
ancillary classes (including type_info) as being visible. So, on gcc, just change;

template<class T>
class Foo : public Bar
{ ... };

to;

template<class T>
class __attribute__((visibility("default"))) Foo : public Bar
{ ... };

This works for GCC on both Linux and Mac.

On windows, I think that you would write something like;

template<class T>
class __declspec(dllexport) Foo : public Bar
{ ... };

though I've not tried this yet.

In Sire, I use the macro SIRE_EXPORT to identify symbols that must be exported (this is copied
from Q_DECL_EXPORT provided by Qt), so in this commit I have changed the declaration of 
all polymorphic template classes to;

template<class T>
class SIRE_EXPORT Foo : public Bar
{ ... };

While this will all work for libraries that are linked with the executable, I am not
sure that this fixes problems when loading libraries as a plugin (e.g. loading a 
library connected to a python module). For plugins to work, I believe that the 
library must be loaded with a dlopen() with RTLD_GLOBAL.