I used shark (provided as part of the developer tools on the mac) to perform a time profile of sire, and found that I lost a lot of time doing;
mol0.isEmpty() or mol1.isEmpty()

This is because mol?.isEmpty() was calling mol.nCutGroups() == 0, which was calling
(partialmolecule).selection(), which was creating a copy of an AtomSelection! Removing this
call and replacing it with a direct count of the CoordGroupArray has significantly sped up
Sire. Isn't profiling great!

See ideas/current-speed for the full details, but now on two cores, Sire is running
at over 1/2 a GFLOP!