I've added configure tests to see if lseek64 is available, and if not, then a stub function is used (thereby removing the symbol unavailable dlopen errors when using mpich2).
I have also fixed the QMutex errors on exit by switching over to using Q_GLOBAL_STATIC 
and making sure that the functions check whether or not the static objects
still exist. The cause of the error was NodeDeleter, which was running a background
thread that was checking a global static mutex. I've moved this global static
mutex so that it is a local mutex in a global NodeDeleter object, and have added
a QThread::wait() in the destructor so the mutex won't be deleted until the 
background NodeDeleter thread has exited (and thus stoped using the mutex).