I've written most of the replica exchange code...
However, I've realised that the design is not right - the information about the move is in the wrong
place and this is not extendable to generic replica exchange moves over temperature, pressure etc.

I've worked out a new design, which will replace pretty much everything that I have just written.

C'est la vie

Code is broken