I've finished propogating the changes in Property throughout the rest of the code, and it all compiles and seems to work (the 'sire' application runs and returns the right energy).
Ok, so I'll admit that this change took longer than I thought (I thought it would be 
about an hour or so to do, but it was more involved than I thought as it involved
renaming lots of important classes, e.g. MoleculeGroup, ForceField etc.).

Code seems to be working, but the Python wrappers need to be regenerated and tested