I've got the backend of the inter-process communication working now, and sire_python is able to see how messages are exchanged between processes. There's an occasional deadlock at shutdown, but it looks like it is working.
I've also added seeding of qrand as otherwise QUuid was producing identical UIDs on different
MPI processes (which was quite annoying!). The fix is to call QUuid first so that it sets
its initial seed, then call qsrand to seed it ourselves, and then subsequent calls to
QUuid will then produce truly unique IDs (well, within the constraints of a pseudo-random
number generator).