I've fixed a few bugs:
MoleculeGroup::add(MoleculeGroup) was just broken, and has been fixed

ProtoMS reader has been fixed to read in more data from ProtoMS, so that I can now
get almost exact agreement in the energy. The energy for a DANA (neuraminidase) inhibitor
in a box of TIP4P is -19237.729 kcal mol-1 in ProtoMS (calculated in 0.8s), while
it is -19237.730131 kcal mol-1 in Sire (calculated in 0.3s).

This is close enough agreement, given two completely different languages, different
order of forming the sum etc.