I've fixed memory problems when running a replica simulation - TrajectoryMonitor was holding an unnecessary copy of the molecule group, and Replica was holding a whole other swapped replica, when it only needed to hold a SimStore - these two things meant that sire_python was taking too much memory when running the protein simulation.
Now, I can run the neuraminidase protein with 8 replicas in about 300-400MB, so can
run two mpi jobs comfortably on my laptop (and additional replicas don't cost
any more memory as they are packed to disk!)

I think that I have (finally!) solved the protein memory problems. The next issue
may be that it just takes too long to serialise and deserialise the system, 
although running 8 replicas on 2 processors took 5 minutes, so it is not a showstopper.