I've fixed a bug in linearap.cpp that meant that the linear assignment problem was rarely incorrectly solved.
I've added a 4D matrix class and have extended the API of the 2D matrix class
to include reflecting the data about diagonals and lines.

I've updated linearp to use the NMatrix class rather than Array2D<double>

I've begun to sort out the code in Squire for the orbitals, and am 
sketching out a design.

I've updated the generation of the python wrappers so that functions with
unknown return policies are automatically not wrapped.

Code is broken