I've added a charge constraint class, which allows the charges to be changed during a simulation (using the same interface as the MoleculeConstraint classes). This will allow for AM1BCC charges to be calculated dynamically as a molecule moves. I've written out the skeleton of the AM1BCC class, but this currently only returns zero charges.
I've also added a kludgy fix that makes AtomProperty<T>::array() work correctly.
The problem was that these classes are exposed as typedefs, and Typedef::array()
failed in the python wrappers with an lvalue error, saying that 
AtomProperty<T>::array( Typedef ) was not a valid argument. The kludge 
solution was to explicitly call these functions when Sire.Mol is loaded,
which seems to tell boost::python that Typedef and AtomProperty<T> are 
indeed the same same class.