I've finished (and tested - see test/Squire/am1bcc.py) the AM1BCC charge generator.
It seems to correctly calculate AM1-BCC charges using antechamber, and includes a z-matrix
type test so that charges only need to be recalculated when the conformation of the 
molecule changes (so preventing recalculation if the molecule is just translated
or rotated)

I've also fixed a small bug in the zmatrix code, whereby a null zmatrix is returned
if one is constructed from a basic zmatrix with a complete molecule. This bug
does not affect any current or previous users.