OK  - this is the start of a very broken set of versions of Sire. I am now starting the migration of the code over to Siren. I assumed I could do this piecemeal, but I can't, so this will all break now.
If you want a non-broken version, please use trunk or branches/devel, which are both
copies of the last version of Sire (and branches/devel will continue to be developed while
I am running production on HPCx)

Also - MAJOR - I am dumping the previous version of the binary streaming format. The
Siren streaming format is too different to support what I had before, and given the
low number of current users, it is too expensive to try to port it. Thus any binary
files produced by previous versions of Sire now will not work with any versions
of Sire from now on. This is a pain, but the benefit is that the new streaming
format has XML, better versioning, and allows easy extension to different formats
without having to change the objects (you just create a new polymorphic streaming
object)

Finally, I am taking the opportunity of removing SireError and SireStream
(as these are replaced by Siren) and of removing the dependency of GSL.
I have nothing against GSL, but since I pull BLAS in myself, I want
to remove the GSL dependency as it compliates the portability of the
build scripts if different BLAS libraries are used for GSL and for SireMaths...

Code is very, very, very broken...!