I'm making progress porting SireCluster to Siren. I've cleaned up the classes used for communication. I now have;
Backend - this provides a compute unit that can process WorkPackets

DormantBackend - this handle holds a Backend that is not yet assigned to a cluster
ActiveBackend - this handle holds a Backend that is assigned to a cluster, and 
                is thus connected to by a Frontend (so it is part of a Nodes cluster)

Frontend - this provides a frontend to a compute unit - it provides the means by 
           which you can communicate with a Backend

DormantFrontend - this handle holds a frontend that is not yet in use, i.e. an
                  idle node in the cluster
ActiveFrontend - this handle holds a frontend in active use, i.e. a Node which is
                 being used by the user and is held by a Node object.

Thus, Nodes holds a collection of DormantFrontend objects. When Nodes is asked to
provide a Node, it activates one of the DormantFrontend objects and passes the
resulting ActiveFrontend to a Node object, which is returned to the user.

The user uses the Node object to submit the workpacket and wait on the promise.
The active frontend in the Node communicates the workpacket to the ActiveBackend
(which may be on a different process or computer), and the active backend then
passes the workpacket to the actual Backend object that runs the job. ActiveBackend,
DormantBackend, ActiveFrontend, DormantFrontend, Node and Nodes are all 
concrete, non-virtual classes. All of the flexibility is provided by Frontend
and Backend, with Frontend providing the different routes by which frontends
can connect to backends (e.g. MPIFrontend, SocketFrontend etc.) and backends
providing the different compute resources (e.g. LocalBackend, ThreadBackend,
GPGPUBackend?)

Code is broken