**IMPORTANT BUGFIX**
I've found a bug in the updating of LJ parameters that can affect the validity
of single topology simulations. This has not affected any production simulations
run so far (as far as I am aware - to test run test/SireMove/test_for_leaks.py
on your .s3 restart file, and it will print errors if you are affected by 
this bug). I've fixed the offending code. I strongly encourage you to update
to this version if you are planning to run single topology simulations.