In celebration of Sire's fifth birthday tomorrow, I'm committing the dump of the original subversion repository used to develop Sire, and a copy of all of the original subversion commit comments.
The code has come a long way in five years!

The dump file is bzipped, and can be loaded using the commands;

bunzip2 orig_sire_repository.dump.bz2
svnadmin create orig_sire_repository
svnadmin load orig_sire_repository < orig_sire_repository.dump

svn co file:///path/to/orig_sire_repository orig_sire

This repository contains the development of Sire from the first line of code
(when Sire was known as ProtoMS3, and was just going to be a re-write of 
ProtoMS2 in C++, on 5 Feb 2005), to the version that was first committed into the
google code repository (version 831 in the original repository, version
2 for the google repository, code from 25 July 2006, committed to google
on 7 August 2006).

ProtoMS3 was renamed to Sire when I moved to Wales, in July 2005, so the
name is not yet 5 years old. However, tomorrow is the code's official
birthday :-)

Put together, across the two repositories, there have been now over
2020 source code commits, thus averaging 400 per year. Admittedly though,
831 of those commits were in the 18 months... Since then, I've only managed
about 300 commits a year...!