I've fixed SignatureLock so that it now works (it can sign and verify using PublicKey and PrivateKey).
I've also updated Lock so that it can encrypt/sign and decrypt/verify files.

The next (last step!) is to tie this together to create a full signing lock which 
compresses, encrypts and then signs data. This will be used in SireCluster to allow
all messages to be compressed, encrypted and signed. This will prevent eavesdropping,
and will allow each node to verify the owner of a workpacket and the integrity of that
workpacket.