I've got the dynamics code working!!! :-)
test/SireMove/dynamics.py runs 10 ps of dynamics of a cube of salt (5x5x5). The
energy is conserved, the dynamics looks good, and things are great!

In being able to do dynamics, Sire has taken a big step beyond just being
a better version of ProtoMS. The API for dynamics is quite cool too, 
as it cleanly separates the dynamics integrator from the MoleculeDynamics
move class, and classes that are used to manage the system. The only down-side
is that the updates to the system could end up being very slow...