I've fixed the second bug found by Julien - this was caused by the operator== code for the GeometryComponent classes incorrectly saying they were equal. This was because of a silly typo, e.g.
bool DistanceComponent::operator==(const DistanceComponent &other) const
{
    return this != &other or ...;
}

should be

    return this == &other or ...;

I've scanned the code to see if I made this error anywhere else, and fortunately
I hadn't. I've fixed this bug in all of the GeometryComponent classes and now
the script works. It is interesting to note however that the time to run
10000 moves of test/SireSystem/testmonitors.py increase from 4.5 to 5.5 seconds
when I add the third monitor. I will performance debug this now...

Note that the code is still broken because of the work in progress for the
QM forcefields.

Code is broken