I've updated the version of Siren in devel to match the latest version from branches/users/chryswoods. The merge command was;
svn merge -r 1110:1323 ../users/chryswoods/src/libs/Siren src/libs/Siren

This is so that the fixes to Siren so that it compiles with Qt 4.2 would
work, and also so that the latest version of Siren is available in 
devel in preparation for me to backport the new SireCluster to devel.

The corelib code all compiles and runs, but this code needs to have
the updated python Siren wrappers.

Code is broken