I've fixed a bug in CovalentBondHunter whereby it never found any bonds because it used a too small covalent radius - it now uses the bond order radius. I've added an anchors example script to show how anchors are supposed to work.
I am also trying now to speed up the code for large numbers of molecules,
in particular getting rid of some unnecessary order-N bits of code which
were seriously destroying the performance of the code for large water
boxes (hence the bigbox.py added example).

While this version compiles and runs, I wouldn't trust it yet - I've
done too much to the core and need to run a lot of tests.