I've changed the "at" function of ChunkedVector to remove bounds checking and have switched the intermolecular forcefields over to using ChunkedVector::at rather than ChunkedVector::operator[]. This has improved the speed of the energy calculations by about 25-30% (see test/SireMove/bigbox.py).
However, I am still performing a distance check between all pairs of 
molecules, which means that I still have an O(n) cost in the 
energy evaluation. This is painful, especially as this is evaluated
via a virtual function call (as it is Space dependent). I need
to think of some way of creating patches and breaking the system
down into patches (to remove the O(n) scaling).