BUGFIX: Fixed a critical bug in Sire that was causing incorrect energy evaluation.
The bug is very specific, and does not affect previously published simulations.
The bug is in InterGroup**FF, and causes the energy of the ith molecule in group
0 and the energy of the ith molecule in group 1 to be neglected. As the majority
of use cases for Sire involve 1 or 2 molecules in group0, and multiple molecules
in group1, this means that only the energy between the group0 molecules and the 
first one or two molecules in group1 are skipped (and these molecules tend to 
be on the edge of the solvent box, so outside of the cutoff of the molecules 
in group0).

This bug has been fixed, and it is recommended that you update your code if
you are using Sire.