I've got the code compiling and working, with the sirentest executable now available to run automatic unit tests of all of the classes :-)
I've fixed bugs in the registration of classes (including a weird bug in which the 
QAtomicPointer that held the class registry was reset to zero between static initialisation
and running the main function - weird! I worked around it by removing the thread-safe
code used for static initialisation).

Code is finally working :-)