A small amount of code to fix a big bug! I implemented the parallel version of TestPacket and came across a weird bug as the code crashed at the end of processing. It ended up being caused by TestPacket not properly returning the next object at the end of processing.
This has revealed that I have a problem in the design due to the WorkQueue not being
able to detect when a WorkPacket has finished, or is still in need of processing. 
I think that the answer is that I need to say that the WorkQueue keeps running 
a WorkPacket while a packet is being returned. I must then separate out the 
statistics / health score associated with a job run from the actual run, so that it
can be returned separately to the queue.

Code compiles, links and sirentest runs as expected.